Introduces a yoast (plugin) helper
