+++ src/surfaces/helpers-surface.php
@@ -35,6 +35,7 @@ use YoastSEO_Vendor\Symfony\Component\DependencyInjection\ContainerInterface;
  * @property Helpers\User_Helper           $user
  * @property Helpers\Woocommerce_Helper    $woocommerce
  * @property Helpers\WordPress_Helper      $wordpress
+ * @property Helpers\Yoast_Helper          $yoast
  */
 class Helpers_Surface {
